Fix SKIP_BUNDLING configuration
SKIP_BUNDLING, when enabled, previously exited the script before the ip.txt resource was created, which would cause a runtime failure since the App would not know what address to use to connect to the React Native Bundler. This was hidden by the fact that if you had previously built the app using another configuration, the ip.txt file would exist. But after a 'clean' of the project folder, it would be deleted and the failure would start to occur.