Add some documentation about cookie handling.
I'm not sure this is the best place for the note, but I've not be able to find any documentation—on Chromium, on Android's CookieManager class, or elsewhere—that the SameSite cookies are not synced. I expect some further documentation on limitations of syncing cookies would be valuable, but I don't have the expertise to flesh that out at the moment.