Give better error feedback and print stacktrace
Just spent way too long trying to trace a very simple problem with `react-native link`

Trying to link my project I kept getting: 

```
rnpm-install ERR! ERRPACKAGEJSON No package found. Are you sure it's a React Native project?

Cannot read property '_text' of undefined
```

Which makes it seem like its a package.json problem, which it was not at all. 

Turns out the problem was I had a comment at the end of my main AndroidManifest.xml and the XML parser was failing because of it. How could I possibly know that from this error message? I had to find this file and print the error which then gave me some hints as to what was going on... This should be the default. I don't think we should trade pretty and useless error messages for actually useful ones.