[Android] Replaced ForwardingSink with OutputStream to fix progress
Using a ForwardingSink, an IllegalStateException was thrown in Okio's
RealBufferedSink when attempting to write to a sink that was closed.
Additionally, it did not send updates for non-input stream request bodies.
Replacing with an OutputStream-based sink prevents the crash by throwing an
IOException instead, and fixes the progress updates. Also, now get the content
length before writing to avoid incorrect total size for input streams.
Addresses issues: 10423/11016.