Keep the ignored files during the upgrade process
When the user adds new files to the .gitignore file,
those files are deleted during the upgrade process
because it starts by creating a fresh .gitignore file
without user's modification. This commit adds a step
to move the excluded file at the repository level so
that they are kept ignored during the upgrade process.