Check Network on Request
This is fixing react-native issue #8615.  The problem was that on initialization the _connectionType variable was still set to its default value of RCTConnectionTypeUnknown.  The exposed API method did nothing to determine this unless a subscription had be established and then the method would simply return the last reported value.  Instead, the exposed oneshot API call now actually checks the connection status through the same methods as the subscription and updates RCTNetInfo’s values before returning.

In order to avoid reporting events without a subscription, a flag is set and unset on calls to startObserving and stopObserving.