Seperated SCNetworkReachabilityRef instances
Previously, the one-time and subscription based APIs relyed on a class level `SCNetworkReachabilityRef` instance.  However, closing after the one-shot caused the instance to no longer be usable.  Now there is a general method, `getReachabilityRef`, to get a new reachability ref.  The one-shot instance is closed directly after use but the subscription instance is stored for reuse in `stopObserving`.