Setting ANDROID_LONG_TIMER_MESSAGE to '(ADVICE)'
This commit is based on a suggestion made by a Firebase member https://github.com/facebook/react-native/issues/17083.
Longer timers are used in critical parts of Firebase-SDK such as refreshing the token; Its a bad user experience if the token expires and users need to keep login to the application! We appreciate your cooperation.