Checking For the GET or HEAD in send method
The following change is suggested by a Firebase member here. https://github.com/firebase/firebase-js-sdk/issues/283
It follows the whatwg.org's send method guidelines. https://xhr.spec.whatwg.org/#dom-xmlhttprequest-send
At the section 3 of 4.5.6, it is suggests to check for the GET or HEAD methods, for every XMLHttpRequest send() method, then set the body to null if it is true.