Improve line-height calculation accuracy.
Round the division by 2 result up and down so that the sum is still the
same as the lineheight when the additional variable happens to be odd.

Added test test to see that bottom-top is still equals the line height
when it's odd. This test does not care which value the implementation
chooses to round up or down as long as the sum is correct.