Build single variant when `deviceId` is given.
Fixes #18409

There's a bug when `deviceId` option is given `./gradlew build` is
executed resulting in *all* variants being built. This is a problem in
projects which have >1 flavor. Only single variant should be built in
this case.

Small hack with `glob` was made to make resolution of `.apk` file
compatible with different versions of gradle plugin:

1. gradle 2 outputs apk to
`[...]/outputs/apk/app-<variant-with-dashes>-debug.apk`
2. gradle 3 outputs apk to
`[...]/outputs/apk/<flavor>/<config>/app-<variant-with-dashes>-debug.apk`

using `glob` it is possible to make it work regardless of version of
gradle plugin.