Fix root view cleanup
The tearDownReactContext method of ReactInstanceManager was not removing
root views from its private mAttachedRootViews member.

The perceived problem was that when an Android application with multiple
root views was reloaded, root JS components that existed prior to that
reload were being remounted. They did not appear on screen, but React
life cycle methods such as componentWillMount were being called.