Fixes UI glitch when drawing selective corners with borderRadius
On devices (not simulator) running iOS 11 there seems to be a bug when drawing selective borders with a corner radius. A "bleeding" line appears on corners that don't have a radius while an adjacent corner has.

It’s most likely a bug with iOS 11 itself as the original code is according to spec and it works fine in an iOS 11 simulator. However it turns out that this patch is a workaround for this problem.

This patch increases the stretchable area to 2pt instead of 1pt after witch the glitch dissappears. I assume shouldn’t have a negative performance impact as it also enforces resizingMode UIImageResizingModeStretch.