rename yoga podspec to React-yoga to prevent clash with YogaKit
Adding `YogaKit` to a brownfield React Native project caused a name clash, because `Yoga`, a dependency of `YogaKit` did install in the same folder as the `yoga` library that comes with React Native, which made it impossible to use `YogaKit` in an React Native app.
By renaming the bundled yoga podspec they no longer install in the same directory.