OkHttp is more strict than other http libraries. It crashes with IllegalStateException in case you pass wrong URL. It crashes if it meets unexpected symbols in the header name and value, while standard says it is not recommneded to use those symbols not that they are prohibited.
The headers handing is a special use case as client might have an auth
tokens in the header. In this case we want to get 401 error response
from the server to find out that token is wrong. In case of the onerror
client will continue to retry with an existing token.

Verify that android app receives an error instead of crashing in case of following network request:
"contacts/v2/users/"
Verify the case with passing the headers with

Release Notes
[ANDROID] [BUGFIX] [Networking] - Passing invalid url not crashes the app instead onerror callback of HttpClient is called. Invalid symbols are stripped from the headers to allow http query to fail with 401 errorcode in case of the broken token.

@MsSourceId: ede64de66f9bb19ff1253eabe872f7d3fd60c98c