Fix regression in Android unlink command
Commit 4dfdec9b28074b311ffd75084d9d2a3cbd41a800 introduced a regression
in the `unlink` command for Android projects: native modules previously
linked with the now-deprecated `compile` configuration will not be
detected anymore and by extension will not be unlinked.

`unlink` should be aware of both new and deprecated options to ensure
functionality also in existing projects that previously linked modules
with the old convention.

This commit fixes the regression by:
- first, `makeDeprecatedNativeModule` is added so to have the deprecated
patch line that could be present in the Gradle build file.
- updating the Android `isInstalled` to check, for a given package
name, for both `compile` and `implementation` dependencies records in
`build.gradle`.
- Android `unregisterNativeModule` will proceed to revoke both
possible patch lines from `build.gradle`.