Fix: Dialog (RN Modal) brings app permanently out of immersive mode
See the following link:
https://stackoverflow.com/a/23207365
for more info.

This does fix the issue, however, another, perhaps related issue
persists, the dialog/modal occasionally uses what would be the height of the
app in non-immersive mode, in immersive mode. Meaning that the background,
what ever it is set as, does not use the full available height.