iOS: Support inline view truncation
If text is truncated and an inline view appears after the truncation point, the user should not see the inline view. Instead, we have a bug such that the inline view is always visible at the end of the visible text.

This commit fixes this by marking the inline view as hidden if it appears after the truncation point.

**Before fix**

Inline view is visible even though it appears after the truncation point:

![image](https://user-images.githubusercontent.com/199935/46381972-7d39d380-c65d-11e8-8354-91f65af651e4.png)

**After fix**

Inline view is properly truncated:

![image](https://user-images.githubusercontent.com/199935/46381918-3ba92880-c65d-11e8-8e61-1d6a1aae2c72.png)

**Test Plan**

Tested that the inline view is hidden if it appears after the truncation point when `numberOfLines` is 1 and 2. Similarly, verified that the inline view is visible if it appears before the truncation point.

Adam Comella
Microsoft Corp.