Use hasOwnProperty to prevent accidental hashProperty inheritance.
This was a fairly serious undocumented spec deviation: before this change,
if an object was ever used as a Map key, any other objects with that
object in their prototype chains would appear to have the same hash,
because the hashProperty key would be inherited.

In my case, I was storing Object.getPrototypeOf(someObject) in a Map as
part of a strategy for determining if I'd ever seen a structurally similar
object before. Imagine the chaos when that prototype object happened to be
none other than... Object.prototype! :skull::jack_o_lantern::ghost: