Comment out invalid Buck ruletypes
Summary:
My previous diff broke the open source build, because it uses a Buck version from July when the fb_native_wrapper list was generated in October. Therefore, any rules added to Buck after that point won't exist and break the build.

This list was generated by doing the following:

```
$ git clone https://github.com/facebook/buck.git ~/buck --branch v2018.07.23.01 --depth=1
$ cd buck && ant
$ bin/buck audit buildruletypes
```

Then comparing the list it gives back with the one currently in fb_native_wrapper.bzl

Reviewed By: TheSavior

Differential Revision: D12878931

fbshipit-source-id: be9a7e84b68e9350b7d8e491d00b8025f6577fd5