Android Text: Fix `letterSpacing` rendering when `fontSize` changes
If you set `letterSpacing` on an outer text and then set `fontSize` on an inner text, the wrong `letterSpacing` will be rendered.

Here's an example:

```
<Text style={{ fontSize: 10, letterSpacing: 10 }}>
  <Text style={{ fontSize: 20 }}>Some text</Text>
</Text>
```

`fontSize` affects letter spacing. In this case, the bug is that setting `fontSize` to `20` doesn't cause the letter spacing to be recalculated.

Notice that the logic for applying letter spacing only applies it if the node has a different value for `getEffectiveLetterSpacing()` than its parent. The problem is that `getEffectiveLetterSpacing()` doesn't represent the final letter spacing value -- it doesn't take `fontSize` into account. Consequently, it sometimes incorrectly skips applying letter spacing as illustrated above.

The fix is to ensure `getEffectiveLetterSpacing()` represents the final rendered letter spacing value. To do this, some of the letter spacing calculation code was moved from `CustomLetterSpacingSpan` to `getEffectiveLetterSpacing()`.

Test Plan:
----------

Created a test app where I set `fontSize` and `letterSpacing` in different nested combinations and verified that letter spacing is rendered correctly. Did this with `Text` and `TextInput`. Also verified that the `TextInput's` `placeholder` works properly with `letterSpacing`.

Changelog:
----------

[Android] [Fixed] - Text: Fix `letterSpacing` rendering when `fontSize` changes