Add allowScrollOutOfBounds to ScrollView
When using the scrollTo() function the user can specify a x/y position
that tells the ScrollView to scroll to. Prior to this patch the x/y were
not validated which could cause the SrollView to scroll to a position
outside the content view bounds, leaving the screen empty (the content
disapears, since it goes outside the screen). In order to fix this problem,
this commit imposes that the x/y must be inside the view bounds if allowScrollOutOfBounds
is set to false. This new prop is also set to true automatically to not break
existing apps which may depends on this feature.

Note that this behavior does not occur on Android, because the
Android's framework itself implements this features via [1] and [2], which are used
by React Native.

[1]: https://developer.android.com/reference/android/widget/ScrollView#smoothScrollTo(int,%20int)
[2]: https://developer.android.com/reference/android/widget/ScrollView.html#scrollTo(int,%20int)