RNTester: Add `allowFontScaling` example
iOS's `allowFontScaling` example was copied to Android. Also, I added a case illustrating that `allowFontScaling` gets inherited in nested `<Text>` scenarios. This tests the bug fix from #22917. This nested case was added to both iOS and Android.

Also, moved the comment in `TextAttributes.java` from the top of the file to the top of the class definiiton as requested in #22917.

Test Plan:
----------

Used the OS settings to increase the font scale to the maximum. Opened RNTester on iOS and Android and verified that the text in the `allowFontScaling` example that has `allowFontScaling` set to `false` does not get scaled.

Changelog:
----------

[Android] [Added] - RNTester: Added demo of `Text's` `allowFontScaling` prop

Adam Comella
Microsoft Corp.