Set rounded rectangle mask on ripples
This commit fixes an issue where ripple touch feedback extends beyond
the border radius of a view.

It achieves this by adding a mask to the RippleDrawable background,
collecting that information from two new methods on ReactViewGroup:

1. getBorderRadiusMask() returns a drawable rounded rectangle matching
the view's border radius properties
2. getBorderRadius() produces a float[] with the border radius
information required to build a RoundedRectShape in
getBorderRadiusMask()

Additionally, this commit updates setBorderRadius in ReactViewManager
to re-apply the background whenever it is set, which is necessary to
update the mask on the RippleDrawable background image as the border
radius changes.

Related issues:
https://github.com/facebook/react-native/issues/6480