Use View.INVISIBLE instead of View.GONE to hide ActivityIndicator
On Android, if the ActivityIndicator component is initially set to
animate={false}, it does not display if later set to true.

For some reason, the layout width/height of the associated ProgressBar
remains 0, despite the parent layout having the correct width/height.

I noticed that changing the visibility settings from View.GONE to
View.INVISIBLE didn't have an impact on the React Native layout, and
doing so appears to alleviate the issue.

Related to:
https://github.com/facebook/react-native/issues/9023